feat: consolidate db schemas defs
- added "drop" script and documentation for how to use these two scripts
